pyproject.toml
[project]
name = "olki-backend"
version = "0.1.0"
description = "Paint landing page backend"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.14.0",
    "psycopg2-binary>=2.9.9",
    "pika>=1.3.2",
    "django-prometheus>=2.3.1",
    "Pillow>=10.0.0",
    "markdown>=3.5.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-django>=4.7.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "coverage>=7.3.0",
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
    "factory-boy>=3.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["olki_backend", "products", "contacts"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"tests.py" = ["ARG001", "ARG002"]  # unused function arguments in tests
"test_*.py" = ["ARG001", "ARG002"]  # unused function arguments in tests
"*_tests.py" = ["ARG001", "ARG002"]  # unused function arguments in tests
"views.py" = ["ARG002"]  # unused args/kwargs in Django views
"management/commands/*.py" = ["ARG001"]  # unused arguments in management commands

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
